refactor(06-02): extract read-contract core logic to read-contract-core.ts
- Create read-contract-core.ts with readContractCore, ReadContractCoreInput, ReadContractResult
- Move stepHandler, getUserIdFromExecution helper, and ReadContractResult type to core file
- Update read-contract.ts to import from core and delegate to readContractCore
- No "use step" directive in core file -- safe to import from any future step
